Display TEXT_OVERLAY over grid cell
A text overlay is applied over each grid cell. This allows to have
people know how to vote for each image: what number to type in, which
iMessage reaction to use…

I've choose a font (Jost) that can be redistributed freely. Its license
is available in the file "FONT LICENSE".