Export frames as a video
A video is generated from a combination of each frame. I prefered to
generate a video instead of a GIF, as videos can keep more colours and
have a smaller filesize.

I choose the FFV1 codec to keep all the details in the (combined) images.
This is especially useful to keep the artists intentions, when combining
artistic images.

The FPS is a temporary value. I plan to replace it with an average taken
from each animated images.